move escape fun to utils
